requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.1.0
pip==23.3.1
pytz==2023.3.post1
setuptools==58.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
